feat(root): re-enable the disable XHC1 (causes wakeup from USB)
There is a problem when activating suspend (sleep) mode, in that the
computer immediately wakes up again. This seems to be caused by the USB
acpi interrupt. So far the only suggestion that seems to work is to
disable the XHC1 wakeup entirely. This does mean that you are not able
to wake up your computer by moving the mouse or pressing keys on the
keyboard, so its not the perfect solution, but it works for now.